notificador/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
configparser==5.2.0
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
six==1.16.0
telepot==12.7
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
yarl==1.7.2
requests==2.31.0
python-dotenv==1.0.1